readme.d
The goal of this peer review project is to learn how to analyze Bluetooth and Wi-Fi networks and use Amazon EC2 cloud services. Using the Bluetooth Scan app, the Bluetooth wireless network status and connections are to be evaluated. Next using the Wi-Fi Network Analyzer app, the device and gateway addresses as well as the data rate, channel, and bandwidth measurements are to be evaluated. Then in the Amazon EC2 project, an AWS cloud is set up and then two EC2 cloud services are recommended to be tested. In your project report, include all details you have observed and experienced.